Moving AutoHeight to shared directory
